// apps/web-evmos/src/graphql/types/general_types.tsx
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ActionCoin: any;
  ActionDelegation: any;
  ActionPagination: any;
  ActionRedelegation: any;
  ActionUnbondingDelegation: any;
  _coin: any;
  _dec_coin: any;
  _text: any;
  bigint: any;
  jsonb: any;
  numeric: any;
  smallint: any;
  timestamp: any;
};
// ...
export type Staking_Params = {
  __typename?: 'staking_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Staking_Pool = {
  __typename?: 'staking_pool';
  bonded_tokens: Scalars['String'];
  height: Scalars['bigint'];
  not_bonded_tokens: Scalars['String'];
};
// ...
export type Supply = {
  __typename?: 'supply';
  coins: Scalars['_coin'];
  height: Scalars['bigint'];
};
// ...
export type TokenomicsQuery = {
  stakingParams: Array<{ __typename?: 'staking_params' } & Pick<Staking_Params, 'params'>>;
  stakingPool: Array<
    { __typename?: 'staking_pool' } & {
      bonded: Staking_Pool['bonded_tokens'];
      unbonded: Staking_Pool['not_bonded_tokens'];
    }
  >;
  supply: Array<{ __typename?: 'supply' } & Pick<Supply, 'coins'>>;
};